requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
itsdangerous==2.1.2
pyotp==2.9.0
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Authlib==1.2.1
requests==2.31.0
